fix: 모듈을 정리함(root module-child module 구조를 띠도록 변경)
