+++ src/auth/auth.controller.ts
@@ -10,14 +10,15 @@ import {
 import { AuthGuard } from '@nestjs/passport';
 import { ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { Response } from 'express';
-import { User } from 'src/users/entities/user.entity';
-import { SignupRequestDto } from '../users/dto/signup-request.dto';
+import { User } from 'src/user-repository/entities/user.entity';
+import { SignupRequestDto } from './dto/signup-request.dto';
 import { UsersService } from '../users/users.service';
 import { AuthService } from './auth.service';
 import { SigninMfaRequestDto } from './dto/signin-mfa-request.dto';
 import { UserSigninResponseDto } from './dto/user-signin-response.dto';
 import { FtAuthService } from './ft-auth.service';
 import { GetUser } from './get-user.decorator';
+import { UsersRepository } from '../user-repository/users.repository';
 
 @Controller('auth')
 @ApiTags('auth')
@@ -25,7 +26,7 @@ export class AuthController {
 	constructor(
 		private readonly authService: AuthService,
 		private readonly ftAuthService: FtAuthService,
-		private readonly usersService: UsersService,
+		private readonly usersRepository: UsersRepository,
 	) {}
 
 	@Post('/signin')
@@ -83,7 +84,7 @@ export class AuthController {
 		const nickname = signupRequestDto.nickname;
 		const avatar = signupRequestDto.avatar;
 
-		await this.usersService.signup(user.id, nickname, avatar);
+		await this.authService.signup(user.id, nickname, avatar);
 	}
 
 	// TODO: 테스트용 코드. 추후 삭제
@@ -100,7 +101,9 @@ export class AuthController {
 	@ApiResponse({ status: 201, description: '토큰 발급 성공' })
 	async testSignIn(@Body('nickname') nickname: string, @Res() res: Response) {
 		// 이미 존재하는 유저인지 확인한다.
-		const existUser = await this.usersService.findUserByNickname(nickname);
+		const existUser = await this.usersRepository.findUserByNickname(
+			nickname,
+		);
 		// 이미 존재하는 유저라면 토큰을 발급한다.
 		if (existUser) {
 			const jwtAccessToken = await this.authService.generateJwtToken(
@@ -131,7 +134,10 @@ export class AuthController {
 			return res.send(userSigninResponseDto);
 		}
 
-		const user = await this.usersService.createUser(nickname, 'test@test');
+		const user = await this.usersRepository.createUser(
+			nickname,
+			'test@test',
+		);
 
 		const jwtAccessToken = await this.authService.generateJwtToken(user);
 
@@ -168,7 +174,7 @@ export class AuthController {
 	})
 	@UseGuards(AuthGuard('access'))
 	async signout(@GetUser() user: User, @Res() res: Response) {
-		await this.usersService.signout(user.id);
+		await this.authService.signout(user.id);
 
 		res.clearCookie('accessToken');
 		res.clearCookie('refreshToken');
