+++ src/auth/auth.service.ts
@@ -1,6 +1,5 @@
 import {
 	BadRequestException,
-	ConflictException,
 	Inject,
 	Injectable,
 	Logger,
@@ -13,11 +12,13 @@ import { Socket } from 'socket.io';
 import * as speakeasy from 'speakeasy';
 import jwtConfig from 'src/config/jwt.config';
 import userConfig from 'src/config/user.config';
-import { User } from 'src/users/entities/user.entity';
-import { UsersRepository } from '../users/users.repository';
+import { User } from 'src/user-repository/entities/user.entity';
+import { UsersRepository } from '../user-repository/users.repository';
 import { FtUserParamDto } from './dto/ft-user-param.dto';
 import { SigninMfaRequestDto } from './dto/signin-mfa-request.dto';
 import { UserFindReturnDto } from './dto/user-find-return.dto';
+import { UserStatus } from '../common/enum';
+import { DBUpdateFailureException } from '../common/exception/custom-exception';
 
 @Injectable()
 export class AuthService {
@@ -102,11 +103,58 @@ export class AuthService {
 		return accessToken;
 	}
 
+	async signup(userId: number, nickname: string, avatar: string) {
+		const user = await this.validateUserExist(userId);
+
+		await this.validateUserAlreadySignUp(user);
+
+		await this.validateNickname(nickname);
+
+		const updateRes = await this.usersRepository.update(userId, {
+			avatar: avatar,
+			nickname: nickname,
+			status: UserStatus.ONLINE,
+		});
+
+		if (updateRes.affected !== 1) {
+			throw DBUpdateFailureException(`user ${userId} update failed`);
+		}
+	}
+
+	async validateUserExist(userId: number) {
+		const user = await this.usersRepository.findOne({
+			where: {
+				id: userId,
+			},
+		});
+
+		if (!user) {
+			throw new BadRequestException(
+				`User with id ${userId} doesn't exist`,
+			);
+		}
+		return user;
+	}
+
+	async validateUserAlreadySignUp(user: User) {
+		if (user.nickname)
+			throw new BadRequestException(`${user.id} is already signed up`);
+	}
+
 	async validateNickname(nickname: string) {
 		const user = await this.usersRepository.findUserByNickname(nickname);
-		if (user) {
-			throw new ConflictException('이미 존재하는 닉네임입니다.');
-		}
+
+		if (user)
+			throw new BadRequestException(
+				`nickname '${nickname}' is already exist! Try again`,
+			);
+	}
+
+	async signout(userId: number) {
+		await this.usersRepository.update(userId, {
+			refreshToken: null,
+			status: UserStatus.OFFLINE,
+		});
 	}
 
 	async getUserFromSocket(client: Socket) {
