+++ src/auth/dto/signup-request.dto.ts
@@ -1,7 +1,6 @@
 import { IsNotEmpty, IsString, Length } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 
-
 export class SignupRequestDto {
 	@ApiProperty({ description: '아바타' })
 	@IsNotEmpty()
