+++ src/auth/dto/user-find-return.dto.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/users/entities/user.entity';
+import { User } from 'src/user-repository/entities/user.entity';
 
 export type UserFindReturnDto = {
 	user: User;
