+++ src/channels/channels.module.ts
@@ -9,14 +9,23 @@ import { ChannelsService } from './channels.service';
 import { ChannelInvitation } from './entities/channel-invitation.entity';
 import { ChannelUser } from './entities/channel-user.entity';
 import { Channel } from './entities/channel.entity';
-import { UsersModule } from '../users/users.module';
-import { AuthModule } from '../auth/auth.module';
+import { JwtModule } from '@nestjs/jwt';
+import jwtConfig from '../config/jwt.config';
+import { ConfigType } from '@nestjs/config';
+import { UserRepositoryModule } from '../user-repository/user-repository.module';
+import { FriendsModule } from '../friends/friends.module';
 
 @Module({
 	imports: [
+		JwtModule.registerAsync({
+			inject: [jwtConfig.KEY],
+			useFactory: (jwtConfigure: ConfigType<typeof jwtConfig>) =>
+				jwtConfigure,
+		}),
 		TypeOrmModule.forFeature([Channel, ChannelUser, ChannelInvitation]),
-		AuthModule,
-		UsersModule,
+		// AuthModule,
+		UserRepositoryModule,
+		FriendsModule,
 	],
 	controllers: [ChannelsController],
 	providers: [
