+++ src/channels/channels.service.ts
@@ -5,8 +5,8 @@ import { Redis } from 'ioredis';
 import { MUTE_TIME } from 'src/common/constants';
 import { ChannelType, ChannelUserType } from 'src/common/enum';
 import { GatewayCreateChannelInvitationParamDto } from 'src/game/dto/gateway-create-channelInvitation-param-dto';
-import { User } from 'src/users/entities/user.entity';
-import { UsersRepository } from 'src/users/users.repository';
+import { User } from 'src/user-repository/entities/user.entity';
+import { UsersRepository } from 'src/user-repository/users.repository';
 import { DBUpdateFailureException } from '../common/exception/custom-exception';
 import { ChannelInvitationRepository } from './channel-invitation.repository';
 import { ChannelUsersRepository } from './channel-users.repository';
