+++ src/friends/blocks.service.spec.ts
