+++ src/friends/dto/channel-socket-id-response.dto.ts
