+++ src/friends/entities/friend.entity.ts
