+++ src/game/dto/accept-game-param.dto.ts
@@ -1,4 +1,4 @@
-import { User } from '../../users/entities/user.entity';
+import { User } from '../../user-repository/entities/user.entity';
 
 export class acceptGameParamDto {
 	invitationId: number;
