+++ src/game/dto/create-invitation-param.dto.ts
@@ -1,6 +1,6 @@
 import { GameType } from '../../common/enum';
 import { IsIn, IsInt, IsPositive, IsString } from 'class-validator';
-import { User } from '../../users/entities/user.entity';
+import { User } from '../../user-repository/entities/user.entity';
 
 export class CreateGameInvitationParamDto {
 	invitingUser: User;
