+++ src/game/game.controller.ts
@@ -11,7 +11,7 @@ import {
 import { GameService } from './game.service';
 import { CreateGameInvitationRequestDto } from './dto/create-invitation-request.dto';
 import { GetUser } from '../auth/get-user.decorator';
-import { User } from '../users/entities/user.entity';
+import { User } from '../user-repository/entities/user.entity';
 import { ApiTags } from '@nestjs/swagger';
 import { CreateGameInvitationParamDto } from './dto/create-invitation-param.dto';
 import { PositiveIntPipe } from '../common/pipes/positiveInt.pipe';
