+++ src/user-repository/dto/user-profile-return.dto.ts
@@ -1,0 +1,12 @@
+export type UserProfileReturnDto = {
+	id: number;
+	nickname: string;
+	avatar: string;
+	statusMessage: string | null;
+	loseCount: number;
+	winCount: number;
+	totalCount: number;
+	ladderScore: number;
+	ladderMaxScore: number;
+	ladderRank: number;
+};
