+++ src/users/dto/user-profile-response.dto.ts
@@ -12,16 +12,3 @@ export type UserProfileResponseDto = {
 	isFriend: boolean;
 	isBlocked: boolean;
 };
-
-export type UserProfileReturnDto = {
-	id: number;
-	nickname: string;
-	avatar: string;
-	statusMessage: string | null;
-	loseCount: number;
-	winCount: number;
-	totalCount: number;
-	ladderScore: number;
-	ladderMaxScore: number;
-	ladderRank: number;
-};
