+++ src/users/users.module.ts
@@ -1,43 +1,21 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Game } from 'src/game/entities/game.entity';
-import { GameRepository } from 'src/game/game.repository';
-import { BlocksRepository } from './blocks.repository';
-import { BlocksService } from './blocks.service';
-import { Block } from './entities/block.entity';
-import { Friend } from './entities/friend.entity';
-import { User } from './entities/user.entity';
-import { FriendsRepository } from './friends.repository';
-import { FriendsService } from './friends.service';
 import { RanksService } from './ranks.service';
 import { UsersController } from './users.controller';
-import { UsersRepository } from './users.repository';
 import { UsersService } from './users.service';
 import { ScheduleModule } from '@nestjs/schedule';
 import { AppService } from 'src/app.service';
+import { UserRepositoryModule } from '../user-repository/user-repository.module';
+import { GameModule } from '../game/game.module';
+import { FriendsModule } from '../friends/friends.module';
 
 @Module({
 	imports: [
-		TypeOrmModule.forFeature([User, Friend, Block, Game]),
 		ScheduleModule.forRoot(),
+		UserRepositoryModule,
+		FriendsModule,
+		GameModule,
 	],
 	controllers: [UsersController],
-	providers: [
-		AppService,
-		RanksService,
-		UsersService,
-		FriendsService,
-		BlocksService,
-		UsersRepository,
-		FriendsRepository,
-		BlocksRepository,
-		GameRepository,
-	],
-	exports: [
-		UsersService,
-		UsersRepository,
-		FriendsRepository,
-		BlocksRepository,
-	],
+	providers: [AppService, RanksService, UsersService],
 })
 export class UsersModule {}
